LibWeb: Use i64 when generating initial counter representation
This means that we can correctly represent the absolute value of INT_MIN
(2147483648) rather than clamping it to an i32 (2147483647)

+++ Libraries/LibWeb/CSS/CounterStyle.h
@@ -36,7 +36,7 @@ public:
     Optional<FlyString> const& fallback() const { return m_fallback; }
     CounterStylePad const& pad() const { return m_pad; }
 
-    Optional<String> generate_an_initial_representation_for_the_counter_value(i32 value) const;
+    Optional<String> generate_an_initial_representation_for_the_counter_value(i64 value) const;
     bool uses_a_negative_sign() const;
 
     virtual ~CounterStyle() = default;
